pyproject.toml
[tool.poetry]
name = "trello-tasks"
version = "0.1.0"
description = ""
authors = ["Alejandro Piad <alepiad@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
py-trello = "^0.16.0"
typer = "^0.1.1"
pyyaml = "^5.3.1"
psutil = "^5.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
coverage = "^5.0.4"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tools.poetry]
packages = [
    { include="trellotasks" }
]
